refactor breadcrumbs into separate component
